+++ src/main/kotlin/com/github/etkachev/nxwebstorm/models/RunSchematicConfig.kt
@@ -1,0 +1,8 @@
+package com.github.etkachev.nxwebstorm.models
+
+data class RunSchematicConfig(
+  val cli: CliCommands,
+  val command: String,
+  val name: String,
+  val args: Map<String, String>
+)