+++ src/main/kotlin/com/github/etkachev/nxwebstorm/models/SchematicActionButtonPlacement.kt
@@ -1,0 +1,9 @@
+package com.github.etkachev.nxwebstorm.models
+
+/**
+ * Placement of where to place the action buttons on the schematic tool window
+ */
+enum class SchematicActionButtonPlacement(val data: String) {
+  TOP("top"),
+  BOTTOM("bottom")
+}