+++ src/main/kotlin/com/github/etkachev/nxwebstorm/ui/RunSchematicPanel.kt
@@ -3,17 +3,24 @@ package com.github.etkachev.nxwebstorm.ui
 import com.github.etkachev.nxwebstorm.actionlisteners.CheckboxListener
 import com.github.etkachev.nxwebstorm.actionlisteners.TextControlListener
 import com.github.etkachev.nxwebstorm.models.FormValueMap
+import com.github.etkachev.nxwebstorm.models.SchematicActionButtonPlacement
+import com.github.etkachev.nxwebstorm.services.MyProjectService
+import com.github.etkachev.nxwebstorm.ui.buttons.SchematicActionButtons
 import com.github.etkachev.nxwebstorm.utils.FormCombo
 import com.github.etkachev.nxwebstorm.utils.FormControlType
 import com.github.etkachev.nxwebstorm.utils.GenerateFormControl
 import com.github.etkachev.nxwebstorm.utils.ReadFile
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
+import com.intellij.openapi.actionSystem.ActionPlaces
+import com.intellij.openapi.actionSystem.ActionToolbar
+import com.intellij.openapi.actionSystem.DefaultActionGroup
+import com.intellij.openapi.actionSystem.Separator
+import com.intellij.openapi.actionSystem.impl.ActionButton
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogPanel
 import com.intellij.ui.components.JBTextArea
 import com.intellij.ui.layout.panel
-import java.awt.event.ActionEvent
 import javax.swing.BorderFactory
 import javax.swing.DefaultComboBoxModel
 import javax.swing.JComponent
@@ -26,9 +33,10 @@ class RunSchematicPanel(
   schematicLocation: String,
   private val formMap: FormValueMap = FormValueMap()
 ) {
-  var json = ReadFile(project).readJsonFromFileUrl(schematicLocation)
+  var json = ReadFile.getInstance(project).readJsonFromFileUrl(schematicLocation)
   var required: JsonArray? = null
   private var formControlGenerator: GenerateFormControl
+  private val nxService = MyProjectService.getInstance(project)
 
   init {
     required = if (json?.has("required") == true) json!!.get("required").asJsonArray else null
@@ -61,29 +69,66 @@ class RunSchematicPanel(
     return formControls
   }
 
+  private fun getActionGroup(
+    runAction: () -> Unit,
+    debugAction: () -> Unit,
+    dryRunAction: () -> Unit
+  ): List<ActionButton> {
+    val actionGroup = DefaultActionGroup()
+    actionGroup.add(SchematicActionButtons.run(runAction))
+    actionGroup.add(SchematicActionButtons.debug(debugAction))
+    actionGroup.add(SchematicActionButtons.dryRun(dryRunAction))
+    val myActions = actionGroup.getChildren(null)
+    val buttonList = mutableListOf<ActionButton>()
+    for (action in myActions) {
+      if (action is Separator) {
+        continue
+      }
+      val presentation = action.templatePresentation
+      val button = ActionButton(action, presentation, ActionPlaces.UNKNOWN, ActionToolbar.DEFAULT_MINIMUM_BUTTON_SIZE)
+      buttonList.add(button)
+    }
+    return buttonList
+  }
+
+  private fun getFormRowData(control: FormCombo): FormRowData {
+    val key = control.name
+    val vg = formMap.valueGetter(key)
+    val vs = formMap.valueSetter(key)
+    val vbg = formMap.boolValueGetter(key)
+    val vbs = formMap.boolValueSetter(key)
+    val nvg = formMap.nullValueGetter(key)
+    val nvs = formMap.nullValueSetter(key)
+    val descriptionLabel = getWrappedTextAreaForLabel(control.description ?: "")
+    return FormRowData(vg, vs, vbg, vbs, nvg, nvs, descriptionLabel)
+  }
+
   fun generateCenterPanel(
     withBorder: Boolean = false,
     addButtons: Boolean = false,
-    dryRunAction: (ActionEvent) -> Unit = {},
-    runAction: (ActionEvent) -> Unit = {}
+    dryRunAction: () -> Unit = {},
+    runAction: () -> Unit = {},
+    debugAction: () -> Unit = {}
   ): JComponent? {
     val props = json?.get("properties")?.asJsonObject ?: return null
     val formControls = getFormControlKeys(props)
+    val actions = this.getActionGroup(runAction, debugAction, dryRunAction)
+    val (runBtn, debugBtn, dryRunBtn) = actions
 
     val panel = panel {
+      if (addButtons && nxService.actionBarPlacement == SchematicActionButtonPlacement.TOP) {
+        row {
+          runBtn()
+          dryRunBtn()
+          debugBtn()
+        }
+      }
       row {
         label("ng generate workspace-schematic:$id")
       }
       formControls.mapNotNull { control ->
         val comp = control.component ?: return@mapNotNull null
-        val key = control.name
-        val vg = formMap.valueGetter(key)
-        val vs = formMap.valueSetter(key)
-        val vbg = formMap.boolValueGetter(key)
-        val vbs = formMap.boolValueSetter(key)
-        val nvg = formMap.nullValueGetter(key)
-        val nvs = formMap.nullValueSetter(key)
-        val descriptionLabel = getWrappedTextAreaForLabel(control.description ?: "")
+        val (vg, vs, vbg, vbs, nvg, nvs, descriptionLabel) = getFormRowData(control)
         titledRow(control.finalName) {
           if (control.type != FormControlType.BOOL) {
             row { descriptionLabel() }
@@ -112,12 +157,11 @@ class RunSchematicPanel(
           }
         }
       }
-      if (addButtons) {
+      if (addButtons && nxService.actionBarPlacement == SchematicActionButtonPlacement.BOTTOM) {
         row {
-          button("Dry Run", dryRunAction)
-          right {
-            button("Run", runAction)
-          }
+          runBtn()
+          dryRunBtn()
+          debugBtn()
         }
       }
     }
@@ -132,3 +176,13 @@ class RunSchematicPanel(
     }
   }
 }
+
+data class FormRowData(
+  val valueGetter: () -> String,
+  val valueSetter: (String) -> Unit,
+  val boolValueGetter: () -> Boolean,
+  val boolValueSetter: (Boolean) -> Unit,
+  val nullValueGetter: () -> String?,
+  val nullValueSetter: (String?) -> Unit,
+  val descriptionLabel: JBTextArea
+)