+++ src/main/kotlin/com/github/etkachev/nxwebstorm/utils/GenerateTerminalCommand.kt
@@ -1,5 +1,6 @@
 package com.github.etkachev.nxwebstorm.utils
 
+import com.github.etkachev.nxwebstorm.models.CliCommands
 import com.github.etkachev.nxwebstorm.models.NxProjectType
 
 fun getSchematicCommandFromValues(
@@ -9,9 +10,19 @@ fun getSchematicCommandFromValues(
   nxProjectType: NxProjectType,
   dryRun: Boolean = true
 ): String {
-  val keys = values.keys
   val dryRunString = if (dryRun) " --dry-run" else ""
-  val flagCommands = keys.mapNotNull { key ->
+  val (nxPath, nxExec) = CliCommands.NX.data
+  val (ngPath, ngExec) = CliCommands.NG.data
+  val nx = "node $nxPath/$nxExec"
+  val ng = "node $ngPath/$ngExec"
+  val prefix = if (type == "workspace-schematic") "$nx workspace-schematic $id" else "$nx generate $type:$id"
+  val finalPrefix = if (nxProjectType == NxProjectType.Nx) prefix else "$ng generate $type:$id"
+  val flags = getCommandArguments(values).joinToString(" ")
+  return "$finalPrefix $flags --no-interactive$dryRunString"
+}
+
+fun getCommandArguments(values: Map<String, String>): List<String> {
+  return values.keys.mapNotNull { key ->
     val value = values[key]
     val finalText = if (value == "true" || value == "false") {
       if (value == "true") "--$key" else null
@@ -23,10 +34,4 @@ fun getSchematicCommandFromValues(
     }
     finalText
   }
-  val nx = "node node_modules/@nrwl/cli/bin/nx.js"
-  val ng = "node node_modules/@angular/cli/bin/ng"
-  val prefix = if (type == "workspace-schematic") "$nx workspace-schematic $id" else "$nx generate $type:$id"
-  val finalPrefix = if (nxProjectType == NxProjectType.Nx) prefix else "$ng generate $type:$id"
-  val flags = flagCommands.joinToString(" ")
-  return "$finalPrefix $flags --no-interactive$dryRunString"
 }